requirements.txt
langchain>=0.2.14
langchain-community>=0.2.12
langchain-openai>=0.1.22
langchain-ollama>=0.1.1
chromadb>=0.5.5
sentence-transformers>=3.0.0
pypdf>=4.2.0
python-dotenv>=1.0.1
fastapi>=0.111.0
uvicorn[standard]>=0.30.0
rich>=13.7.1

# 语音处理依赖
pyaudio>=0.2.11
numpy>=1.21.0
openai-whisper>=20231117
SpeechRecognition>=3.10.0
edge-tts>=6.1.0
TTS>=0.22.0
pyttsx3>=2.90
